+++ Gores/PlagueDroplet.cs
@@ -38,7 +38,7 @@ namespace OurStuffAddon.Gores
 				{
 					frameDuration = 24 + Main.rand.Next(96);
 				}
-				if ((int)gore.frameCounter >= frameDuration)
+				if (gore.frameCounter >= frameDuration)
 				{
 					gore.frameCounter = 0;
 					gore.frame += 1;
@@ -53,7 +53,7 @@ namespace OurStuffAddon.Gores
 			else if (gore.frame <= 6)
 			{
 				frameDuration = 8;
-				if ((int)gore.frameCounter >= frameDuration)
+				if (gore.frameCounter >= frameDuration)
 				{
 					gore.frameCounter = 0;
 					gore.frame += 1;
@@ -75,7 +75,7 @@ namespace OurStuffAddon.Gores
 				{
 					gore.velocity.Y = 12f;
 				}
-				if ((int)gore.frameCounter >= frameDuration)
+				if (gore.frameCounter >= frameDuration)
 				{
 					gore.frameCounter = 0;
 					gore.frame += 1;
@@ -88,7 +88,7 @@ namespace OurStuffAddon.Gores
 			else
 			{
 				gore.velocity.Y += 0.1f;
-				if ((int)gore.frameCounter >= frameDuration)
+				if (gore.frameCounter >= frameDuration)
 				{
 					gore.frameCounter = 0;
 					gore.frame += 1;
@@ -124,7 +124,7 @@ namespace OurStuffAddon.Gores
 				if (Main.tile[tileX, tileY] != null && Main.tile[tileX, tileY].liquid > 0)
 				{
 					gore.velocity *= 0f;
-					gore.position.Y = tileY * 16 - (int)(Main.tile[tileX, tileY].liquid / 16);
+					gore.position.Y = tileY * 16 - Main.tile[tileX, tileY].liquid / 16;
 				}
 			}
