+++ NPCs/Bosses/AncientObserver/AncientPebbleShot.cs
@@ -28,7 +28,7 @@ namespace OurStuffAddon.NPCs.Bosses.AncientObserver
 		public override void AI()
 		{
 			projectile.ai[0] += 1f;
-			projectile.rotation = (float)Math.Atan2((double)projectile.velocity.Y, (double)projectile.velocity.X) + 1.57f;
+			projectile.rotation = (float)Math.Atan2(projectile.velocity.Y, projectile.velocity.X) + 1.57f;
 			projectile.localAI[0] += 1f;
 			if (projectile.ai[0] >= 300f)       //how much time the projectile can travel before landing
 			{
@@ -36,7 +36,7 @@ namespace OurStuffAddon.NPCs.Bosses.AncientObserver
 				projectile.Kill();
 			}
 			{
-				projectile.rotation = (float)Math.Atan2((double)projectile.velocity.Y, (double)projectile.velocity.X) + 1.57f;
+				projectile.rotation = (float)Math.Atan2(projectile.velocity.Y, projectile.velocity.X) + 1.57f;
 				if (projectile.localAI[0] == 0f)
 				{
 					Main.PlaySound(SoundID.Item20, projectile.position);
@@ -53,8 +53,8 @@ namespace OurStuffAddon.NPCs.Bosses.AncientObserver
 				{
 					Main.dust[dust].velocity += projectile.velocity * 0.2f;
 				}
-				Main.dust[dust].position.X = projectile.Center.X + 4f + (float)Main.rand.Next(-2, 3);
-				Main.dust[dust].position.Y = projectile.Center.Y + (float)Main.rand.Next(-2, 3);
+				Main.dust[dust].position.X = projectile.Center.X + 4f + Main.rand.Next(-2, 3);
+				Main.dust[dust].position.Y = projectile.Center.Y + Main.rand.Next(-2, 3);
 				Main.dust[dust].noGravity = true;
 			}
 		}
