+++ Projectiles/CosmicLaser.cs
@@ -29,7 +29,7 @@ namespace OurStuffAddon.Projectiles
 
 		public override void AI()
 		{
-			projectile.rotation = (float)Math.Atan2((double)projectile.velocity.Y, (double)projectile.velocity.X) + .785f;
+			projectile.rotation = (float)Math.Atan2(projectile.velocity.Y, projectile.velocity.X) + .785f;
 		}
 	}
 }
